% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendaR.R
\name{calendaR}
\alias{calendaR}
\title{ggplot2 monthly and yearly calendars}
\usage{
calendaR(
  start = c("S", "M"),
  month = NULL,
  year = format(Sys.Date(), "\%Y"),
  weeknames = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday",
    "Sunday"),
  position = c("horizontal", "vertical"),
  title,
  title.size = 20,
  title.col = "gray30",
  motivation = "",
  motivation.col = "gray30",
  text = "",
  text.at = NULL,
  text.size = 4,
  text.style = "plain",
  text.col = "gray30",
  special.days = NULL,
  special.col = rgb(0, 0, 1, alpha = 0.25),
  gradient = FALSE,
  col = "gray30",
  lwd = 0.75,
  lty = 1,
  font.family = "sans",
  weekdays.col = "gray30",
  month.col = "gray30",
  days.col = "gray30",
  day.size = 3,
  pdf = FALSE
)
}
\arguments{
\item{start}{\code{"S"} (default) for starting the week on Sunday or \code{"M"} for starting the week on Monday.}

\item{month}{Month of the year or NULL for the yearly calendar.}

\item{year}{Year of the calendar. By default uses the current year.}

\item{weeknames}{Names of the days of the week. By default they will be in english.}

\item{position}{\code{"vertical"} or \code{"horizontal"}}

\item{title}{Title of the the calendar. If not supplied is the year and the month, or the year if \code{month = NULL}.}

\item{title.size}{Size of the main title.}

\item{title.col}{Color of the main title.}

\item{motivation}{Motivational phrase added as subtitle of the plot.}

\item{motivation.col}{Color of the motivational phrase.}

\item{text}{Character vector of texts to be added on the calendar. Only for monthly calendars.}

\item{text.at}{Number of days where to add the texts of the \code{text} argument.}

\item{text.size}{Font size of the texts added with the \code{text} argument.}

\item{text.style}{Style of the texts added with the \code{text} argument.}

\item{text.col}{Color of the texts added with the \code{text} argument.}

\item{special.days}{Numeric vector indicating the days to color or \code{"weekend"} for coloring all the weekends.}

\item{special.col}{Color for the days indicated in special.days.}

\item{gradient}{Boolean. If \code{special.days} is a numeric vector of the length of the displayed days, \code{gradient = TRUE} creates a gradient of the \code{special.col} on the calendar.}

\item{col}{Color of the lines of the calendar.}

\item{lwd}{Line width of the calendar.}

\item{lty}{Line type of the calendar.}

\item{font.family}{Font family of all the texts except the motivational phrase.}

\item{weekdays.col}{Color of the names of the days.}

\item{month.col}{If \code{month = NULL}, is the color of the month names.}

\item{days.col}{Color of the number of the days.}

\item{day.size}{Font size of the number of the days.}

\item{pdf}{Boolean. If TRUE, saves the calendar in the working directory in A4.}
}
\description{
Create ready to print monthly and yearly calendars with ggplot2. The package allows personalizing colors (even setting a gradient color scale for a full month or year), texts and fonts and adding texts on the days for monthly calendars.
}
\examples{
\dontrun{
# Create ready to print monthly calendars for all the months of the current year
# with week starting on Sunday
invisible(sapply(1:12 , function(i) calendaR(month = i, pdf = T)))
}
}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
