% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendR-package.R
\docType{package}
\name{calendR-package}
\alias{calendR-package}
\alias{_PACKAGE}
\title{Package calendR}
\description{
This package creates monthly and yearly calendars based on ggplot2 package. The function provided allows coloring the days, adding texts, customizing the font colors, styles and fonts and saving ready to print calendars to your working directory in PDF (in landscape or portrait A4 format).
}
\details{
\itemize{
\item{Package: calendR}
\item{Version: 1.0}
\item{Maintainer: José Carlos Soage González \email{jsoage@uvigo.es}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-coder.com/}
}

}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Pérez Veiga, Natalia.}
}
}
