% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendR-package.R
\docType{package}
\name{calendR-package}
\alias{calendR-package}
\title{calendR: Ready to Print Monthly and Yearly Calendars Made with 'ggplot2'}
\description{
This package creates monthly and yearly calendars based on 'ggplot2' package. The function provided allows coloring the days, adding texts, customizing the font colors, styles and fonts and saving ready to print calendars to the working directory in PDF (in landscape or portrait A4 format).
}
\details{
\itemize{
\item{Package: calendR}
\item{Version: 1.0}
\item{Maintainer: José Carlos Soage González \email{jsoage@uvigo.es}}
}
}
\seealso{
\itemize{
\item{\href{https://r-coder.com/}{R programming tutorials}}
\item{\href{https://r-coder.com/calendar-plot-r/}{Tutorial: calendar plots in R with calendR package}}
}
}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Pérez Veiga, Natalia.}
}
}
